chore: split android provider responsibilities
