+++ examples/v0.81.1/src/screens/scenario/native-e2e/androidRequestOptionsResults.ts
@@ -1,0 +1,24 @@
+import type { SetScenarioResult } from "../hooks/useScenarioResults";
+import type { ScenarioResult } from "../types";
+import { createScenarioResults } from "../utils/results";
+
+export const androidRequestOptionsResults = createScenarioResults([
+  "autoProvider",
+  "playServicesProvider",
+  "platformProvider",
+  "fused",
+  "oneShotDistance",
+  "coarseCache",
+  "mixedWatch",
+  "maxUpdates",
+  "headingUnwatch",
+  "invalid",
+  "fineDenied"
+] as const);
+
+export type AndroidRequestOptionsResultKey =
+  keyof typeof androidRequestOptionsResults;
+
+export type AndroidRequestOptionsSetResult = SetScenarioResult<
+  Record<AndroidRequestOptionsResultKey, ScenarioResult>
+>;
